keep track of running commands; use shell langs for shell command
require `run` handlers to return `Maybe[Pid]` and store it as a
`RunningCommand` in the state.
when running a shell command, check whether its shell is already
running. Until now, the shell command would have been executed again (in
the shell).

when parsing a shell command, add the shell's `langs` to the command's
`langs`.